test(vite): refactor tests by feature
